iam_service/src/test/java/org/path/iam/manager/DistrictManagerTest.java
package org.path.iam.manager;

import org.path.iam.dao.DistrictDao;
import org.path.iam.model.District;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import java.util.ArrayList;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.*;

@ExtendWith(MockitoExtension.class)
class DistrictManagerTest {

    @Mock
    private DistrictDao districtDao;

    @InjectMocks
    private DistrictManager districtManager;

    @BeforeEach
    void setUp() {
        // You can initialize any mock behavior here
    }

    @Test
    void testFindAllByStateId() {
        // Prepare test data
        Long stateId = 1L;
        String search = "district";
        Integer pageNumber = 1;
        Integer pageSize = 10;
        List<District> expectedDistricts = List.of(new District(), new District());
        when(districtDao.findAllByStateId(stateId, search, pageNumber, pageSize)).thenReturn(expectedDistricts);

        // Call the method to be tested
        List<District> actualDistricts = districtManager.findAllByStateId(stateId, search, pageNumber, pageSize);

        // Verify the result
        assertEquals(expectedDistricts, actualDistricts);
    }

    @Test
    void testFindAllByStateGeoId() {
        // Prepare test data
        String stateGeoId = "XYZ";
        List<District> expectedDistricts = List.of(new District(), new District());
        when(districtDao.findAllByStateGeoId(stateGeoId)).thenReturn(expectedDistricts);

        // Call the method to be tested
        List<District> actualDistricts = districtManager.findAllByStateGeoId(stateGeoId);

        // Verify the result
        assertEquals(expectedDistricts, actualDistricts);
    }

    @Test
    void testFindAllByGeoIds() {
        // Mock data
        List<District> districts = new ArrayList<>();
        districts.add(new District());
        districts.add(new District());

        // Mock DAO response
        when(districtDao.findAllByGeoIds(anyList())).thenReturn(districts);

        // Perform test
        List<District> result = districtManager.findAllByGeoIds(List.of("geoId1", "geoId2"));

        // Verify result
        assertEquals(2, result.size());
    }

    @Test
    void testGetCountByStateId() {
        // Mock DAO response
        when(districtDao.getCountByStateId(anyLong(), any())).thenReturn(5L);

        // Perform test
        Long result = districtManager.getCountByStateId(1L, "search");

        // Verify result
        assertEquals(5L, result);
    }

    @Test
    void testFindByStateGeoId() {
        // Mock data
        List<District> districts = new ArrayList<>();
        districts.add(new District());
        districts.add(new District());

        // Mock DAO response
        when(districtDao.findByStateGeoId(any())).thenReturn(districts);

        // Perform test
        List<District> result = districtManager.findByStateGeoId("geoId");

        // Verify result
        assertEquals(2, result.size());
    }

    @Test
    void testFindByStateNameAndDistrictName() {
        // Mock data
        District expectedDistrict = new District();

        // Mock DAO response
        when(districtDao.findByStateNameAndDistrictName(anyString(), anyString())).thenReturn(expectedDistrict);

        // Perform test
        District result = districtManager.findByStateNameAndDistrictName("districtName", "stateName");

        // Verify result
        assertEquals(expectedDistrict, result);
    }

    @Test
    void testFindByNameAndStateId() {
        // Mock data
        District expectedDistrict = new District();

        // Mock DAO response
        when(districtDao.findByNameAndStateId(anyString(), anyLong())).thenReturn(expectedDistrict);

        // Perform test
        District result = districtManager.findByNameAndStateId("districtName", 1L);

        // Verify result
        assertEquals(expectedDistrict, result);
    }

    @Test
    void testFindByName() {
        // Mock data
        District expectedDistrict = new District();

        // Mock DAO response
        when(districtDao.findByName(anyString())).thenReturn(expectedDistrict);

        // Perform test
        District result = districtManager.findByName("districtName");

        // Verify result
        assertEquals(expectedDistrict, result);
    }
    // Add more test cases for other methods as needed
}
